fix time convert

+++ src/ontime/static/main.js
@@ -5,8 +5,9 @@ $(function() {
         var time = $t.attr('time');
         if (! time) return;
         var t = time.split(/[:\s-]/);
-        time = Date.UTC(parseInt(t[0]), parseInt(t[1]) - 1, parseInt(t[2]),
-                        parseInt(t[3]), parseInt(t[4]), parseInt(t[5]));
+        time = Date.UTC(
+            parseFloat(t[0]), parseFloat(t[1]) - 1, parseFloat(t[2]),
+            parseFloat(t[3]), parseFloat(t[4]), parseFloat(t[5]));
         time = new Date(time);
         $t.text(OT.dt.formatDateTime(time, true));
     });
